apps/bff/requirements.txt
azure-functions>=1.20.0,<2.0.0
